- Adding Testing on DTOs - Implementing a new algorithm for both diagonal lines

+++ src/main/java/com/mercadolibre/challenge/service/impl/MutantServiceImpl.java
@@ -53,7 +53,7 @@ public class MutantServiceImpl implements MutantService {
 			ratio = BigDecimal.ZERO;
 		}
 
-		return new StatsDto(mutantCount, humanCount, ratio);
+		return new StatsDto(mutantCount, humanCount, ratio.doubleValue());
 	}
 
 }
